api/requirements.txt
isort== 5.13.2
black== 24.8.0
pylint== 3.2.6
coverage== 7.6.1
django== 4.2.16
python-dotenv== 1.0.1
psycopg2-binary== 2.9.9
djangorestframework== 3.15.2
djangorestframework-simplejwt== 5.3.1
